requirements.txt
Django==2.1.1
pip3==18.0
djangorestframework==3.8.2
coreapi==1.9.0
django-filter
djoser